Listener/requirements.txt
mysql-connector-python
selenium
uiautomation
requests
flask
cryptography